feat:chatbot

+++ src/main/java/com/brnsmrt/africanet/repository/ProductRepository.java
@@ -1,14 +1,14 @@
 package com.brnsmrt.africanet.repository;
 
 import com.brnsmrt.africanet.domain.Product;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
-    List<Product> findByBasePriceLessThanEqual(Double maxPrice);
+    List<Product> findByBasePriceLessThanEqual(BigDecimal maxPrice);
     List<Product> findByNameContainingIgnoreCase(String keyword);
     List<Product> findByCategoryId(Long categoryId);
     List<Product> findByCategoryIdIn(List<Long> categoryIds);
